Load image.
This took a while to actually get working. Turns out you need to be
careful with crate versions with Piston. In this case specifying the
version of gfx_device_gl was required. It may be best to pin some of
the other crates though I'll leave it for now unless I hit any other
weird issues.